+++ lib/advance_math.dart
@@ -11,8 +11,9 @@ export 'src/math/geometry/geometry.dart';
 export 'src/math/algebra/algebra.dart';
 export 'src/quantity/quantity.dart';
 export 'src/interpolate/interpolate.dart';
-export 'src/num_words/num_words.dart';
+export 'src/code_translator/num_words/num_words.dart';
 export 'src/code_translator/morse_code.dart';
+export 'src/models/data_frame.dart';
 
 part 'src/utils/num_ext.dart';
 part 'src/utils/string_ext.dart';