+++ CHANGELOG.md
@@ -1,37 +1,63 @@
+## 3.3.8
+
+* **[BUG_FIX]** CoordinateType not set to UTM
+* **[BUG_FIX]** Change the data types of mean and correlation to num
+* **[BROKEN]** NumOrWords has been moved into code translators with the same name as Morse code.
+* **[BROKEN]** MorseCodeTranslator class has been renamed to MorseCode
+* **[FEATURE]** Added `Dataframe` class for working with dataframes or tables.
+* **[IMPROVEMENT]** Removed duplicate functions
+* **[IMPROVEMENT]** Enhanced `combinations` and `permutations` function:
+  * The function now generates the actual combinations or permutations of elements, mimicking the behavior of the combinations function in R.
+  * Added support for Lists as input for `n`.
+  * Introduced optional `func` and simplify parameters for applying functions to combinations or permutations and controlling output structure.
+  * Improved documentation with clearer explanations, examples, and comments.
+* **[IMPROVEMENT]** Cleaned code for expressions (symbolic math)
+* **[IMPROVEMENT]** Added doc-strings to some functions
+* **[IMPROVEMENT]** Added documentation of Morse code in the ReadMe file
+* **[IMPROVEMENT]** Fixed README
+
 ## 3.3.7
+
 * Fixed README
 * Added functions
 
 ## 3.3.6
+
 * Fixed bugs
-* Renamed Row Column, and Diagonal matrices to RowMatrix, ColumnMatrix and DiagonalMatrix repectively
+* Renamed Row, Column, and Diagonal matrices to RowMatrix, ColumnMatrix and DiagonalMatrix repectively
 
 ## 3.3.5
+
 * Added morse code
 * Added more math functions
 * Improved documentation and performance
 * Fixed bugs and aligned code
 
 ## 3.3.4
+
 * Added support for converting number to words
 * Added test for roman numerals.
 * Fixed bugs in Roman numerals
 * Added some string extension (e.g. capitalization, removeSpecialCharacters etc)
 
 ## 3.3.3
+
 * Setting support for expressions
 * Fixed errors in pow for both real and Complex numbers
 
 ## 3.3.2
+
 * Added some math constants
 * Fixed error in angles
 * Fixed error in parsing linear and constant polynomial strings
 * Increased the SDK
 
 ## 3.3.1
+
 * Added Interpolation
 
 ## 3.3.0
+
 * Added ZScore computation
 * Added more functions matrix statistics
 * Added multiple supports for determinant
@@ -39,21 +65,26 @@
 * Fixed documentation inconsistencies
 
 ## 3.2.4
+
 * Fixed error in angles
 
 ## 3.2.3
+
 * Fixed error in angles conversion
 
 ## 3.2.2
+
 * Improved the checks on roman numerals
 * Added conversion between roman numerals and dates
 * Improved the flexibility in arithmetic with integer, roman strings and roman numbers
 * Fixed README file
 
 ## 3.2.1
+
 * Improved the checks on roman numerals
 
 ## 3.2.0
+
 * Added roman numerals
 * Added trigonometry functions supports to work with complex numbers
 * Added limit
@@ -61,6 +92,7 @@
 * Fixed README file
 
 ## 3.1.0
+
 * Simplified Example file
 * Improved exponents of Complex numbers
 * Fixed bearingTo() in Point to include axis
@@ -68,6 +100,7 @@
 * Fixed README file
 
 ## 3.0.0
+
 * Added Geometry (Point, line, Circle, Triangle, Polygon, etc.)
 * Added Polynomial (Linear, Quadratic, Cubic, Quartic, Durand-Kerner)
 * Added List and vecctors having same properties
@@ -75,26 +108,32 @@
 * Fixed README file
 
 ## 2.1.2
+
 * Fixed complex number outputing wrong string
 
 ## 2.1.1
+
 * Fixed bugs
 
 ## 2.1.0
+
 * Added arguments for diagonal 
 * Improved min, max and sum functions with axes
 * Fixed bug with magic matrix not working for singly even numbers (6, 10, 14, 18, 22 etc).
 * Fixed bugs
 
 ## 2.0.2
+
 * Changed the reverse function to flip in matrices
 * Added magic() to the Maatrix constructors
 * Fixed bug in FileIO
 
 ## 2.0.1
+
 * Fixed bug
 
 ## 2.0.0
+
 * Fixed the usage of quantities in the respective computations
 * Implemented support for quantity
 * Fixed Web not supported
@@ -102,23 +141,28 @@
 * Fixed README
 
 ## 1.0.3
+
 * Fixed bugs
 * Fixed README
 
 ## 1.0.2
+
 * Conversion of angles(radians, degrees, gradients, DMS, and DM) 
 * Organized code
 * Fixed bugs
 * Fixed README
 
 ## 1.0.0
+
 * Moved codes and reorganized functions
 * Fixed bugs 
 
 ## 0.1.8
+
 * Fixed bugs (in null space)
 
 ## 0.1.7
+
 * Added rescale for both vectors and matrix
 * Improved vector compatibility with lists
 * Added operations on vectors such as expo, sum, prod, etc.
@@ -127,11 +171,13 @@
 * Fixed bugs
 
 ## 0.1.5
+
 * Added support for distance calculation for vectors and matrices
 * Improved consistency in linear algebra
 * Added scale for vector types
 
 ## 0.1.4
+
 * Spercial matrices and vectors their functionalities
 * Fixed spellings in matrix structure properties
 * Added functions partioning of vectors subVector() and getVector()
@@ -139,14 +185,17 @@
 * Fixed README
 
 ## 0.1.2
+
 * Fixed spellings in matrix structure properties
 * Added Vectors, Complex Numbers, and Complex Vectors to README
 * Fixed README
 
 ## 0.1.1
+
 * Fixed README
 
 ## 0.1.0
+
 * Improved indexOf() and random functionalities
 * Fixed README 
 * Fixed bugs